finora_ml/requirements.txt
pydantic>=2.0
transformers>=4.34.0
sentence-transformers>=2.2.2
chromadb>=0.4.15
google-generativeai>=0.3.0
torch>=2.0.0
python-dotenv>=1.0.0
